% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cvode_calc_derivs}
\alias{cvode_calc_derivs}
\title{Allows calling the model that calculates the time derivatives}
\usage{
cvode_calc_derivs(model_, t, states, parameters, forcings_data_)
}
\description{
Allows calling the model that calculates the time derivatives
}

