% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cvode_calc_jac}
\alias{cvode_calc_jac}
\title{Allows calling the function to calculate the Jacobian matrix of the model}
\usage{
cvode_calc_jac(jacobian_, t, states, parameters, forcings_data_)
}
\description{
Allows calling the function to calculate the Jacobian matrix of the model
}

