% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cvode_Cpp}
\alias{cvode_Cpp}
\title{Simulates the model when it is written as a C++ function}
\usage{
cvode_Cpp(times, states, parameters, forcings_data, settings, model_)
}
\description{
Simulates the model when it is written as a C++ function
}

