% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cvode_R}
\alias{cvode_R}
\title{Simulates the model when it is written as a R function}
\usage{
cvode_R(times, states, parameters, forcings_data, settings, model, jacobian)
}
\description{
Simulates the model when it is written as a R function
}

