% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{example_jacobian}
\alias{example_jacobian}
\title{An example of a model's Jacobian matrix (unit test)}
\usage{
example_jacobian(t, states, parameters, forcings)
}
\description{
An example of a model's Jacobian matrix (unit test)
}

