% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ida_calc_res}
\alias{ida_calc_res}
\title{Allows calling the model that calculates the time derivatives}
\usage{
ida_calc_res(model_, t, states, derivatives, parameters, forcings_data_)
}
\description{
Allows calling the model that calculates the time derivatives
}

