% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ida_Cpp_stl}
\alias{ida_Cpp_stl}
\title{Simulates the model when it is written as a C++ function using stl containers}
\usage{
ida_Cpp_stl(times, states_, derivatives_, parameters_, forcings_data_, settings,
  model_, jacobian_)
}
\description{
Simulates the model when it is written as a C++ function using stl containers
}

