% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wrap_cvodes}
\alias{wrap_cvodes}
\title{Returns the solution to the IVP problem at specified timepoints. 
Object returned is a matrix where nrow = number timepoints &
ncol = no. states + no. observed + [(no. states)x(no. parameters)]
the third term above is only returned when sensitivity equations are being
calculated}
\usage{
wrap_cvodes(times, states_, parameters_, forcings_data_, settings, model_,
  jacobian_)
}
\description{
Returns the solution to the IVP problem at specified timepoints. 
Object returned is a matrix where nrow = number timepoints &
ncol = no. states + no. observed + [(no. states)x(no. parameters)]
the third term above is only returned when sensitivity equations are being
calculated
}

